split out introspection logic
